Fix app bar buttons
