+++ app/src/main/java/uk/ac/cam/cl/juliet/fragments/DataFragment.java
@@ -1,11 +1,9 @@
 package uk.ac.cam.cl.juliet.fragments;
 
 import android.Manifest;
-import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.pm.PackageManager;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
@@ -13,28 +11,11 @@ import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.CheckBox;
 import android.widget.TextView;
-import android.widget.Toast;
-import com.microsoft.graph.concurrency.ICallback;
-import com.microsoft.graph.core.ClientException;
-import com.microsoft.graph.extensions.DriveItem;
-import com.microsoft.identity.client.AuthenticationResult;
-import com.microsoft.identity.client.MsalClientException;
-import com.microsoft.identity.client.MsalException;
-import com.microsoft.identity.client.PublicClientApplication;
-import com.microsoft.identity.client.User;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import uk.ac.cam.cl.juliet.R;
@@ -42,9 +23,6 @@ import uk.ac.cam.cl.juliet.activities.MainActivity;
 import uk.ac.cam.cl.juliet.adapters.FilesListAdapter;
 import uk.ac.cam.cl.juliet.computationengine.Burst;
 import uk.ac.cam.cl.juliet.computationengine.InvalidBurstException;
-import uk.ac.cam.cl.juliet.data.AuthenticationManager;
-import uk.ac.cam.cl.juliet.data.GraphServiceController;
-import uk.ac.cam.cl.juliet.data.IAuthenticationCallback;
 import uk.ac.cam.cl.juliet.data.InternalDataHandler;
 import uk.ac.cam.cl.juliet.models.SingleOrManyBursts;
 
@@ -54,9 +32,7 @@ import uk.ac.cam.cl.juliet.models.SingleOrManyBursts;
  * @author Ben Cole
  */
 public class DataFragment extends Fragment
-        implements FilesListAdapter.OnDataFileSelectedListener,
-                IAuthenticationCallback,
-                MainActivity.PermissionListener {
+        implements FilesListAdapter.OnDataFileSelectedListener, MainActivity.PermissionListener {
 
     public static String TOP_LEVEL = "top_level";
     public static String FILES_LIST = "files_list";
@@ -64,12 +40,9 @@ public class DataFragment extends Fragment
     private RecyclerView filesList;
     private TextView noFilesToDisplayText;
     private FilesListAdapter adapter;
-    private MenuItem signIn;
-    private MenuItem signOut;
     private List<SingleOrManyBursts> files;
-    private User user;
 
-    OnInnerFolderClickedListener listener;
+    DataFragmentListener listener;
 
     /**
      * If this is the fragment displaying the top level then it will load its files globally.
@@ -81,7 +54,6 @@ public class DataFragment extends Fragment
     public View onCreateView(
             @NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_data, container, false);
-        setHasOptionsMenu(true);
 
         // Get the context
         Context context = getContext();
@@ -115,7 +87,7 @@ public class DataFragment extends Fragment
 
         // Subscribe for permission updates
         MainActivity main = (MainActivity) getActivity();
-        main.addListener(this);
+        if (main != null) main.addListener(this);
 
         // Return the View that was created
         return view;
@@ -151,53 +123,6 @@ public class DataFragment extends Fragment
             }
         }
         return new ArrayList<>();
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.sync_button:
-                showSyncDialog();
-                return true;
-            case R.id.sign_in_button:
-                // Handling Microsoft connection
-                connect();
-                return true;
-            case R.id.sign_out_button:
-                // Disconnect
-                // TODO: Display some kind of "signed out" message
-                try {
-                    AuthenticationManager.getInstance().disconnect();
-                } catch (MsalClientException msal) {
-                    msal.printStackTrace();
-                }
-                signOut.setVisible(false);
-                signIn.setVisible(true);
-        }
-        return false;
-    }
-
-    /**
-     * A method that is called on tab selection - checking for a user still logged in
-     *
-     * @param isVisibleToUser
-     */
-    @Override
-    public void setUserVisibleHint(boolean isVisibleToUser) {
-        super.setUserVisibleHint(isVisibleToUser);
-        // Handle viewing the correct menu buttons
-        displayCorrectAuthButtons();
-    }
-
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        menu.clear();
-        inflater.inflate(R.menu.menu_sync, menu);
-        // Init the menu items
-        signIn = menu.getItem(0);
-        signOut = menu.getItem(1);
-        displayCorrectAuthButtons();
-        super.onCreateOptionsMenu(menu, inflater);
     }
 
     /**
@@ -323,53 +248,6 @@ public class DataFragment extends Fragment
         return result;
     }
 
-    /** Displays a dialog for syncing the files with the server. */
-    private void showSyncDialog() {
-        Context context = getContext();
-        if (context == null) return;
-        final Dialog dialog = new Dialog(context);
-        dialog.setContentView(R.layout.dialog_upload_files);
-        final CheckBox deleteAfterUploadingCheckbox =
-                dialog.findViewById(R.id.deleteAfterUploadingCheckbox);
-        dialog.findViewById(R.id.uploadButton)
-                .setOnClickListener(
-                        new Button.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                uploadAllUnsyncedFiles(deleteAfterUploadingCheckbox.isChecked());
-                                dialog.cancel();
-                            }
-                        });
-        dialog.findViewById(R.id.cancelButton)
-                .setOnClickListener(
-                        new Button.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                dialog.cancel();
-                            }
-                        });
-        dialog.show();
-    }
-
-    /**
-     * A method for checking the current authentication status and setting the correct sign in or
-     * out buttons
-     */
-    private void displayCorrectAuthButtons() {
-        if (getView() == null || signIn == null || signOut == null) return;
-        try {
-            if (AuthenticationManager.getInstance().getPublicClient().getUsers().size() == 0) {
-                signIn.setVisible(true);
-                signOut.setVisible(false);
-            } else {
-                signIn.setVisible(false);
-                signOut.setVisible(true);
-            }
-        } catch (MsalClientException msal) {
-            msal.printStackTrace();
-        }
-    }
-
     /** Shows a dialog message to confirm whether a file or folder should be deleted. */
     private void showConfirmDeleteDialog(SingleOrManyBursts file) {
         Context context = getContext();
@@ -402,7 +280,7 @@ public class DataFragment extends Fragment
      *
      * <p>Call when the set of files has been modified.
      */
-    private void notifyFilesChanged() {
+    public void notifyFilesChanged() {
         adapter.notifyDataSetChanged();
     }
 
@@ -417,72 +295,8 @@ public class DataFragment extends Fragment
      */
     private void uploadFile(
             SingleOrManyBursts file, FilesListAdapter.FilesListViewHolder viewHolder) {
-        new UploadFileTask(this, viewHolder).execute(file);
-    }
-
-    /**
-     * Uploads all unsynced files to OneDrive.
-     *
-     * @param deleteAfterUploading true if files should be deleted after uploading; false to keep
-     *     files on device after uploading
-     */
-    private void uploadAllUnsyncedFiles(boolean deleteAfterUploading) {
-        // TODO: implement
-    }
-
-    /** Begins the authentication process with Microsoft */
-    private void connect() {
-        // Get the Authentication Manager Instance
-        AuthenticationManager authManager = AuthenticationManager.getInstance();
-
-        // Get the public client application
-        PublicClientApplication clientApp = authManager.getPublicClient();
-
-        // Try and access the users
-        List<User> users = null;
-
-        try {
-            users = clientApp.getUsers();
-            if (users != null && users.size() == 1) {
-                // There is a cached user so silently login
-                authManager.acquireTokenSilently(users.get(0), true, this);
-            } else {
-                // There are no cached users so interactively login
-                authManager.acquireToken(getActivity(), this);
-            }
-        } catch (MsalClientException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * On successful authentication set the user
-     *
-     * @param res the authetnication result
-     */
-    @Override
-    public void onSuccess(AuthenticationResult res) {
-        user = res.getUser();
-        // Swap visibility of the buttons
-        signIn.setVisible(false);
-        signOut.setVisible(true);
-    }
-
-    /**
-     * Notify if there is an error
-     *
-     * @param msalException
-     */
-    @Override
-    public void onError(MsalException msalException) {
-        Toast.makeText(getContext(), "An error occurred whilst logging you in", Toast.LENGTH_LONG)
-                .show();
-    }
-
-    /** Notify if the user cancels */
-    @Override
-    public void onCancel() {
-        Toast.makeText(getContext(), "The user cancelled logging in", Toast.LENGTH_LONG).show();
+        if (listener == null) return;
+        listener.uploadFile(this, viewHolder, file);
     }
 
     /** Called on permission granted - refresh file listing */
@@ -502,82 +316,13 @@ public class DataFragment extends Fragment
         adapter.notifyDataSetChanged();
     }
 
-    /** Asynchronously uploads a file to OneDrive. */
-    private static class UploadFileTask extends AsyncTask<SingleOrManyBursts, Void, Boolean> {
-
-        private SingleOrManyBursts file;
-        private DataFragment parent;
-        private FilesListAdapter.FilesListViewHolder viewHolder;
-        private GraphServiceController gsc;
-
-        public UploadFileTask(
-                DataFragment parent, FilesListAdapter.FilesListViewHolder viewHolder) {
-            super();
-            this.parent = parent;
-            this.viewHolder = viewHolder;
-            this.gsc = new GraphServiceController();
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            viewHolder.setSpinnerVisibility(true);
-            viewHolder.setSyncStatusVisibility(false);
-        }
-
-        @Override
-        protected Boolean doInBackground(SingleOrManyBursts... files) {
-            if (files.length < 1) return false;
-            try {
-                file = files[0];
-                // Send the data using the graph service controller
-                AuthenticationManager auth = AuthenticationManager.getInstance();
-                InternalDataHandler idh = InternalDataHandler.getInstance();
-                if (auth.isUserLoggedIn()) {
-                    File datafile = idh.getSelectedDataFile();
-                    gsc.uploadDatafile(
-                            file.getNameToDisplay(),
-                            "dat",
-                            idh.convertToBytes(datafile),
-                            new ICallback<DriveItem>() {
-                                @Override
-                                public void success(DriveItem driveItem) {
-                                    Log.d("UPLOAD", "Upload was successful!");
-                                }
-
-                                @Override
-                                public void failure(ClientException ex) {
-                                    ex.printStackTrace();
-                                }
-                            });
-                }
-            } catch (MsalClientException msal) {
-                msal.printStackTrace();
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
-            } catch (IOException io) {
-                io.printStackTrace();
-            }
-            return true;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean success) {
-            super.onPostExecute(success);
-            file.setSyncStatus(success);
-            viewHolder.setSpinnerVisibility(false);
-            viewHolder.setSyncStatusVisibility(true);
-            parent.notifyFilesChanged();
-        }
-    }
-
     /**
-     * Sets hte listener for when a folder is clicked.
+     * Sets the listener for when a folder is clicked.
      *
      * @param listener The listener that will handle displaying the inner folder in place of this
      *     fragment
      */
-    public void setOnInnerFolderClickedListener(OnInnerFolderClickedListener listener) {
+    public void setDataFragmentListener(DataFragmentListener listener) {
         this.listener = listener;
     }
 
@@ -585,7 +330,12 @@ public class DataFragment extends Fragment
      * Used by a wrapper class so that this instance can be replaced with another instance to
      * display the contents of the folder that was selected.
      */
-    public interface OnInnerFolderClickedListener {
+    public interface DataFragmentListener {
         void onInnerFolderClicked(SingleOrManyBursts innerFolder);
+
+        void uploadFile(
+                DataFragment parent,
+                FilesListAdapter.FilesListViewHolder viewHolder,
+                SingleOrManyBursts file);
     }
 }
